fix: homepage redirect on submit
