+++ src/components/allRecipesView.js
@@ -1,31 +1,13 @@
-import React from 'react';
-import Album from './Album'
-import axios from 'axios'
-import {Route, Switch} from 'react-router-dom'
-import SingleRecipeView from './singleRecipeView';
+import React from "react";
+import RecipeCard from "./RecipeCard";
 
 export default class AllRecipesView extends React.Component {
-    constructor(props) {
-        super(props)
-        this.state = {
-            email: '',
-            password: ''
-        }
-        this.handleChange = this.handleChange.bind(this)
-        this.handleSubmit = this.handleSubmit.bind(this)
-    }
+  constructor(props) {
+    super(props);
+    this.state = this.props.location.state;
+  }
 
-    handleChange(evt) {
-        this.setState({ [evt.target.name]: evt.target.value })
-    }
-
-    async handleSubmit(evt) {
-        evt.preventDefault()
-    }
-
-    render() {
-        return (
-            <Album onChange={this.handleChange} onSubmit={this.handleSubmit} state={this.state} />
-        )
-    }
+  render() {
+    return <RecipeCard />;
+  }
 }
