+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
-import { Router } from "react-router-dom";
+import { BrowserRouter as Router } from "react-router-dom";
 import history from "./history";
 
 ReactDOM.render(
